Add files via upload
In this model I will use the example car, Opel Astra. We will define additional derived classes for various parts of the Opel Astra car.
I have added detailed classes for the Opel Astra car. The Wheel class, for example, includes properties such as material and size, 
while the Brake class includes the brake type. The Transmission class includes the transmission type.

When you run the main() function, it will create an instance of the OpelAstra class (myOpelAstra) and call the printDetails() 
function to display the detailed information about the Opel Astra, including the engine, wheels, brakes, and transmission.